add info about editing

+++ CS/BlazorGridUnboundColumns/Data/WeatherForecast.cs
@@ -2,6 +2,6 @@ namespace BlazorGridUnboundColumns.Data {
     public class WeatherForecast {
         public DateTime Date { get; set; }
 
-        public int TemperatureC { get; set; }
+        public double TemperatureC { get; set; }
     }
 }
